Fix template library bug
